requirements.txt
apache-libcloud==2.0.0
appdirs==1.4.3
colorama==0.3.9
Django==1.11
django-admin-sortable2==0.6.11
django-braces==1.11.0
django-constance==2.0.0
django-crispy-forms==1.6.1
django-mail-templated==2.6.5
django-mptt==0.8.7
django-multiupload==0.5.2
django-picklefield==0.3.2
django-polymorphic==1.2
django-polymorphic-tree==1.4
django-tag-parser==3.1
djangorestframework==3.6.2
docutils==0.13.1
future==0.16.0
inflection==0.3.1
lockfile==0.12.2
netifaces==0.10.4
packaging==16.8
psutil==5.1.3
pycrypto==2.6.1
pyparsing==2.2.0
python-daemon==2.1.1
pytz==2017.2
radical.utils==0.45
requests==2.13.0
saga-python==0.45.1
six==1.10.0
waves-adaptors==0.1.0
